- Split Discord Connector into sub-services and cron tasks for easier micro-control - Start adding Flask server for optionally triggering bot
